move documented bodies and schemas to patches folder
