+++ projects/optic/src/commands/capture/patches/patchers/shapes/diff.ts
@@ -5,7 +5,8 @@ import { Ono } from '@jsdevtools/ono';
 import { Result, Ok, Err } from 'ts-results';
 import { JsonPath } from '@useoptic/openapi-io';
 import { OpenAPIV3 } from '@useoptic/openapi-utilities';
-import { Body, SchemaObject } from '../../../../oas/shapes';
+import { SchemaObject } from './schema';
+import { Body } from './documented-bodies';
 import { additionalPropertiesDiffs } from './handlers/additionalProperties';
 import { oneOfKeywordDiffs } from './handlers/oneOf';
 import { requiredKeywordDiffs } from './handlers/required';
