+++ projects/optic/src/commands/capture/patches/patchers/shapes/handlers/additionalProperties.ts
@@ -1,5 +1,6 @@
 import { SupportedOpenAPIVersions } from '@useoptic/openapi-io';
-import { Schema, SchemaObject, ShapeLocation } from '../../../../../oas/shapes';
+import { Schema, SchemaObject } from '../schema';
+import { ShapeLocation } from '../documented-bodies';
 import {
   JsonSchemaKnownKeyword,
   ErrorObject,
