--- projects/optic/src/commands/oas/shapes/index.ts
@@ -1,15 +1,0 @@
-export type { Body } from './body';
-export { DocumentedBody } from './body';
-import { ShapeLocation } from './body';
-
-export { observeBodyShape } from './observations';
-
-export type { ShapeLocation };
-
-export { DocumentedBodies } from './streams/documented-bodies';
-
-export { ShapePatch } from './patches';
-export { ShapePatches } from './streams/patches';
-
-export type { SchemaObject } from './schema';
-export { Schema } from './schema';
