+++ projects/optic/src/commands/oas/specs/index.ts
@@ -1,14 +1,10 @@
 export { readDeferencedSpec } from './io';
 
-import { ShapeLocation } from '../shapes';
-
 export {
   OpenAPIV3,
   isFactVariant,
   OpenApiKind as FactVariants,
 } from '@useoptic/openapi-utilities';
-
-export type SpecLocation = ShapeLocation; // TODO: add union of all conceptual locations
 
 // files
 export { SpecFile, SpecFilesSourcemap } from './files';
